MAINT: Renamed `block_matrix.py` to `block_linear_system.py`
