+++ src/pp_solvers/__init__.py
@@ -1,13 +1,13 @@
 version = "0.0.1"
 
-from .block_matrix import *
+from .block_linear_system import *
 from .solver_mixin import *
 from .mat_utils import *
 from .preconditioners import *
 from .petsc_utils import *
 
 __all__ = []
-__all__.extend(block_matrix.__all__)
+__all__.extend(block_linear_system.__all__)
 __all__.extend(mat_utils.__all__)
 __all__.extend(solver_mixin.__all__)
 __all__.extend(preconditioners.__all__)
