+++ tests/test_options_parsers.py
@@ -16,7 +16,7 @@ from petsc4py import PETSc
 from porepy.applications.test_utils.models import add_mixin
 
 import pp_solvers
-from pp_solvers.block_matrix import BlockLinearSystem
+from pp_solvers.block_linear_system import BlockLinearSystem
 from pp_solvers.dof_manager import DofManager
 from pp_solvers.options_parsers import (
     LinearTransformedScheme,
